+++ api/lib/hyperion/util.rb
@@ -51,6 +51,5 @@ module Hyperion
       end
 
     end
-
   end
 end